feat(cruds): implementando backend dos novos cruds(atividade, crianca, plano e registro)

+++ projeto_aps_back/src/main/java/br/edu/ifce/projetoapsback/model/dto/AtividadeSummaryDto.java
@@ -2,7 +2,7 @@ package br.edu.ifce.projetoapsback.model.dto;
 
 import br.edu.ifce.projetoapsback.model.Atividade;
 
-public record AtividadeSummaryDto(Long id, String titulo) {
+public record AtividadeSummaryDto(Integer id, String titulo) {
     public AtividadeSummaryDto(Atividade atividade) {
         this(atividade.getId(), atividade.getTitulo());
     }
