+++ projeto_aps_back/src/main/java/br/edu/ifce/projetoapsback/model/dto/CriancaSummaryDto.java
@@ -2,7 +2,7 @@ package br.edu.ifce.projetoapsback.model.dto;
 
 import br.edu.ifce.projetoapsback.model.Crianca;
 
-public record CriancaSummaryDto(Long id, String nomeCompleto) {
+public record CriancaSummaryDto(Integer id, String nomeCompleto) {
     public CriancaSummaryDto(Crianca crianca) {
         this(crianca.getId(), crianca.getNomeCompleto());
     }
