+++ projeto_aps_back/src/main/java/br/edu/ifce/projetoapsback/model/dto/UserSummaryDto.java
@@ -2,7 +2,7 @@ package br.edu.ifce.projetoapsback.model.dto;
 
 import br.edu.ifce.projetoapsback.model.User;
 
-public record UserSummaryDto(Long id, String fullName, String email) {
+public record UserSummaryDto(Integer id, String fullName, String email) {
     public UserSummaryDto(User user) {
         this(user.getId(), user.getFullName(), user.getEmail());
     }
